+++ sources/src/types.ts
@@ -1,6 +1,0 @@
-import { Entity } from "@loopback/repository";
-
-/** Model Ctor type */
-export type Ctor<Model extends Entity> = typeof Entity & {
-    prototype: Model;
-};