+++ CHANGELOG.md
@@ -1,4 +1,25 @@
-# Changelog
+# [1.2.0](https://github.com/loopback4/loopback-component-filter/compare/1.1.0...1.2.0) (2021-09-12)
+
+
+### Features
+
+* add new Github Action based CI/CD pipelines ([aa7d00c](https://github.com/loopback4/loopback-component-filter/commit/aa7d00cc5526b3e6381bec4ce4f54150d71c5c4e))
+
+## v1.1.0
+
+-   **Feat**: add `context` to options and pass into `InvocationContext`
+
+## v1.0.0
+
+-   **Fix**: remove sources folder from ci file
+
+## v1.0.0
+
+-   **Deps**: update packages
+-   **Clean**: remove `sources` folder
+-   **Feat**: add `model mapper` method for `create`, `update` operations
+-   **Feat**: add unit-tests for `create`, `find`, `edit`, `delete`
+-   **Docs**: add shields to `README`
 
 ## v0.5.0